feature() timelinefeature were you can see articles based on the xml file thats being loaded

+++ backend/src/main/java/com/linkextractor/backend/respositories/LabelRepository.java
@@ -7,7 +7,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 import java.util.Optional;
 
-public interface LabelRepository extends JpaRepository<Label, Long> {
+public interface LabelRepository extends JpaRepository<Label, Integer> {
 
     List<Label> findAll();
